+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Update `PasswordInput` with styled-components.
 - Feature: Add `title` prop on `CheckedCircleIcon`.
 - Feature: Add `avatarBadge` prop on `CommentAvatar`, `Comment` and `CommentForm`.